refactor: move video code out from data generator
this has to be available without evs explorer
